Move common feature code to separate provider
