+++ lib/cisco_node_utils/vni.rb
@@ -33,6 +33,7 @@
 # ----
 #
 require_relative 'node_util'
+require_relative 'feature'
 
 module Cisco
   # node_utils class for Vni
@@ -67,28 +68,9 @@ module Cisco
     end
 
     def self.feature_vni_enable
-      feature_nv_overlay_enable unless feature_nv_overlay_enabled?
+      Feaure.feature_nv_overlay_enable unless
+        Feature.feature_nv_overlay_enabled?
       config_set('vni', 'feature')
-    end
-
-    def self.feature_nv_overlay_supported?
-      config_set('feature', 'nv_overlay')
-    rescue Cisco::CliError => e
-      raise unless e.clierror =~ /not capable of supporting nv overlay feature/
-      false
-    end
-
-    # feature nv overlay
-    def self.feature_nv_overlay_enabled?
-      config_get('feature', 'nv_overlay')
-    rescue Cisco::CliError => e
-      # cmd will syntax reject when feature is not enabled
-      raise unless e.clierror =~ /Syntax error/
-      return false
-    end
-
-    def self.feature_nv_overlay_enable
-      config_set('feature', 'nv_overlay')
     end
 
     def self.mt_full_support
