+++ tests/test_vxlan_vtep.rb
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 require_relative 'ciscotest'
+require_relative '../lib/cisco_node_utils/feature'
 require_relative '../lib/cisco_node_utils/vxlan_vtep'
 require_relative '../lib/cisco_node_utils/vdc'
 
@@ -25,7 +26,7 @@ class TestVxlanVtep < CiscoTestCase
     skip('Platform does not support MT-full or MT-lite') unless
       VxlanVtep.mt_full_support || VxlanVtep.mt_lite_support
     skip('Platform does not support nv overlay feature') unless
-      VxlanVtep.feature_nv_overlay_supported?
+      Feature.feature_nv_overlay_supported?
   end
 
   def teardown
